Move sample MessagesActivity to ui/channel/ChannelActivity
