+++ stream-chat-android-compose-sample/src/main/AndroidManifest.xml
@@ -81,7 +81,7 @@
             android:exported="false"
             />
         <activity
-            android:name=".ui.MessagesActivity"
+            android:name=".ui.channel.ChannelActivity"
             android:exported="false"
             android:windowSoftInputMode="adjustResize"
             />
