+++ stream-chat-android-compose-sample/src/main/java/io/getstream/chat/android/compose/sample/feature/channel/add/AddChannelActivity.kt
@@ -27,7 +27,7 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import io.getstream.chat.android.compose.sample.R
 import io.getstream.chat.android.compose.sample.feature.channel.add.group.AddGroupChannelActivity
-import io.getstream.chat.android.compose.sample.ui.MessagesActivity
+import io.getstream.chat.android.compose.sample.ui.channel.ChannelActivity
 import io.getstream.chat.android.compose.ui.theme.ChatComponentFactory
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
 import io.getstream.chat.android.compose.ui.theme.MessageComposerLeadingContentParams
@@ -82,7 +82,7 @@ class AddChannelActivity : ComponentActivity() {
     }
 
     private fun openChannel(cid: String) {
-        val intent = MessagesActivity.createIntent(this, cid)
+        val intent = ChannelActivity.createIntent(this, cid)
         startActivity(intent)
         finish()
     }
