+++ stream-chat-android-compose-sample/src/main/java/io/getstream/chat/android/compose/sample/feature/channel/list/ChannelsActivity.kt
@@ -66,8 +66,8 @@ import io.getstream.chat.android.compose.sample.feature.channel.add.AddChannelAc
 import io.getstream.chat.android.compose.sample.feature.channel.add.group.AddGroupChannelActivity
 import io.getstream.chat.android.compose.sample.feature.channel.isGroupChannel
 import io.getstream.chat.android.compose.sample.feature.reminders.MessageRemindersActivity
-import io.getstream.chat.android.compose.sample.ui.MessagesActivity
 import io.getstream.chat.android.compose.sample.ui.SampleChatTheme
+import io.getstream.chat.android.compose.sample.ui.channel.ChannelActivity
 import io.getstream.chat.android.compose.sample.ui.channel.DirectChannelInfoActivity
 import io.getstream.chat.android.compose.sample.ui.channel.GroupChannelInfoActivity
 import io.getstream.chat.android.compose.sample.ui.component.AppBottomBar
@@ -359,7 +359,7 @@ class ChannelsActivity : ComponentActivity() {
 
     private fun openMessages(channel: Channel) {
         startActivity(
-            MessagesActivity.createIntent(
+            ChannelActivity.createIntent(
                 context = this,
                 channelId = channel.cid,
                 messageId = null,
@@ -370,7 +370,7 @@ class ChannelsActivity : ComponentActivity() {
 
     private fun openMessages(message: Message) {
         startActivity(
-            MessagesActivity.createIntent(
+            ChannelActivity.createIntent(
                 context = this,
                 channelId = message.cid,
                 messageId = message.id,
@@ -381,7 +381,7 @@ class ChannelsActivity : ComponentActivity() {
 
     private fun openThread(thread: Thread) {
         startActivity(
-            MessagesActivity.createIntent(
+            ChannelActivity.createIntent(
                 context = this,
                 channelId = thread.parentMessage.cid,
                 parentMessageId = thread.parentMessageId,
